Refactor init to follow newer style
